moved flagset classes to model, polish in input_validator and storage helper
