+++ tests/client/test_config.py
@@ -69,29 +69,3 @@ class ConfigSanitizationTests(object):
         processed = config.sanitize('some', configs)
 
         assert processed['redisLocalCacheEnabled']  # check default is True
-
-    def test_sanitize_flag_sets(self):
-        """Test sanitization for flag sets."""
-        flag_sets = config.sanitize_flag_sets([' set1', 'set2 ', 'set3'])
-        assert flag_sets == ['set1', 'set2', 'set3']
-
-        flag_sets = config.sanitize_flag_sets(['1set', '_set2'])
-        assert flag_sets == ['1set']
-
-        flag_sets = config.sanitize_flag_sets(['Set1', 'SET2'])
-        assert flag_sets == ['set1', 'set2']
-
-        flag_sets = config.sanitize_flag_sets(['se\t1', 's/et2', 's*et3', 's!et4', 'se@t5', 'se#t5', 'se$t5', 'se^t5', 'se%t5', 'se&t5'])
-        assert flag_sets == []
-
-        flag_sets = config.sanitize_flag_sets(['set4', 'set1', 'set3', 'set1'])
-        assert flag_sets == ['set1', 'set3', 'set4']
-
-        flag_sets = config.sanitize_flag_sets(['w' * 50, 's' * 51])
-        assert flag_sets == ['w' * 50]
-
-        flag_sets = config.sanitize_flag_sets('set1')
-        assert flag_sets == []
-
-        flag_sets = config.sanitize_flag_sets([12, 33])
-        assert flag_sets == []
