+++ tests/integration/files/splitChanges.json
@@ -58,7 +58,8 @@
                         }
                     ]
                 }
-            ]
+            ],
+            "sets": ["set1", "set2"]
         },
         {
             "orgId": null,
@@ -95,7 +96,8 @@
                         }
                     ]
                 }
-            ]
+            ],
+            "sets": ["set4"]
         },
         {
             "orgId": null,
@@ -136,7 +138,8 @@
                         }
                     ]
                 }
-            ]
+            ],
+            "sets": ["set3"]
         },
         {
             "orgId": null,
@@ -199,7 +202,8 @@
                         }
                     ]
                 }
-            ]
+            ],
+            "sets": ["set1"]
         },
         {
             "orgId": null,
