+++ tests/util/test_storage_helper.py
@@ -1,8 +1,10 @@
 """Storage Helper tests."""
+import pytest
 
 from splitio.util.storage_helper import update_feature_flag_storage, get_valid_flag_sets, combine_valid_flag_sets
 from splitio.storage.inmemmory import InMemorySplitStorage
 from splitio.models import splits
+from splitio.models.flag_sets import FlagSetsFilter
 from tests.sync.test_splits_synchronizer import splits as split_sample
 
 class StorageHelperTests(object):
@@ -24,16 +26,38 @@ class StorageHelperTests(object):
             return False
         storage.is_flag_set_exist = is_flag_set_exist
 
-        storage.config_flag_sets_used = 0
+        class flag_set_filter():
+            def should_filter():
+                return False
+            def intersect(sets):
+                return True
+        storage.flag_set_filter = flag_set_filter
+        storage.flag_set_filter.flag_sets = {}
+
         update_feature_flag_storage(storage, [split], 123)
         assert self.added[0] == split
         assert self.deleted == []
         assert self.change_number == 123
 
-        storage.config_flag_sets_used = 2
+        class flag_set_filter2():
+            def should_filter():
+                return True
+            def intersect(sets):
+                return False
+        storage.flag_set_filter = flag_set_filter2
+        storage.flag_set_filter.flag_sets = set({'set1', 'set2'})
+
         update_feature_flag_storage(storage, [split], 123)
         assert self.added == []
         assert self.deleted[0] == split.name
+
+        class flag_set_filter3():
+            def should_filter():
+                return True
+            def intersect(sets):
+                return True
+        storage.flag_set_filter = flag_set_filter3
+        storage.flag_set_filter.flag_sets = set({'set1', 'set2'})
 
         def is_flag_set_exist2(flag_set):
             return True
@@ -76,22 +100,22 @@ class StorageHelperTests(object):
 
     def test_get_valid_flag_sets(self):
         flag_sets = ['set1', 'set2']
-        config_flag_sets = []
+        config_flag_sets = FlagSetsFilter([])
         assert get_valid_flag_sets(flag_sets, config_flag_sets) == ['set1', 'set2']
 
-        config_flag_sets = ['set1']
+        config_flag_sets = FlagSetsFilter(['set1'])
         assert get_valid_flag_sets(flag_sets, config_flag_sets) == ['set1']
 
         flag_sets = ['set2', 'set3']
-        config_flag_sets = ['set1', 'set2']
+        config_flag_sets = FlagSetsFilter(['set1', 'set2'])
         assert get_valid_flag_sets(flag_sets, config_flag_sets) == ['set2']
 
         flag_sets = ['set3', 'set4']
-        config_flag_sets = ['set1', 'set2']
+        config_flag_sets = FlagSetsFilter(['set1', 'set2'])
         assert get_valid_flag_sets(flag_sets, config_flag_sets) == []
 
         flag_sets = []
-        config_flag_sets = ['set1', 'set2']
+        config_flag_sets = FlagSetsFilter(['set1', 'set2'])
         assert get_valid_flag_sets(flag_sets, config_flag_sets) == []
 
     def test_combine_valid_flag_sets(self):
